added Page4Grade content back in order to retain code
