+++ public/css/Page4Grade.css
@@ -1,0 +1,160 @@
+/* Nutrition Progress Table Styles */
+.nutrition-progress-container {
+    overflow-x: auto;
+    overflow-y: auto;
+    border-radius: 10px;
+    width: 100%;
+    margin-bottom: 20px;
+}
+
+.nutrition-summary-title {
+    position: relative;
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    margin: auto;
+    align-items: left;
+    justify-content: left;
+}
+
+.get-grade-button {
+    position: absolute;
+    top: 35%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    text-decoration: none;
+    background: #8e44ad;
+    color: white;
+    border: none;
+    border-radius: 6px;
+    padding: 10px 15px;
+    cursor: pointer;
+    font-weight: 600;
+    transition: all 0.3s ease;
+    gap: 8px;
+    width: 110px;
+    margin: auto;
+}
+
+.get-grade-button:hover {
+    background: purple;
+    transform: translate(-50%, -54%);
+}
+
+.summary-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    gap: 15px;
+    margin-top: 15px;
+}
+
+.summary-card {
+    background: #f8f9fa;
+    border-radius: 8px;
+    padding: 15px;
+    text-align: center;
+    border-top: 4px solid;
+}
+
+.summary-value {
+    font-size: 1.5rem;
+    font-weight: 700;
+    margin: 8px 0;
+}
+
+.summary-label {
+    color: #7f8c8d;
+    font-size: 0.8rem;
+}
+
+.calories-card {
+    border-top-color: var(--danger-color);
+}
+
+.fats-card {
+    border-top-color: #f39c12;
+}
+
+.cholesterol-card {
+    border-top-color: black;
+}
+
+.sodium-card {
+    border-top-color: #9b59b6;
+}
+
+.carbs-card {
+    border-top-color: var(--success-color);
+}
+
+.fiber-card {
+    border-top-color: #1abc9c;
+}
+
+.sugars-card {
+    border-top-color: black;
+}
+
+.protein-card {
+    border-top-color: var(--primary-color);
+}
+
+.comments {
+    color: black;
+}
+
+.points-card {
+    border-top-color: #8e44ad;
+}
+
+.points-display {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-bottom: 15px;
+}
+
+.points-circle {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    background: #8e44ad;
+    color: white;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-weight: bold;
+    font-size: 2rem;
+    margin-bottom: 12px;
+    box-shadow: 0 3px 10px rgba(0, 0, 0, 0.15);
+}
+
+.points-label {
+    font-size: 1.3rem;
+    font-weight: bold;
+    margin-bottom: 5px;
+}
+
+.points-description {
+    font-size: 0.8rem;
+    color: #7f8c8d;
+    text-align: center;
+    margin-bottom: 15px;
+}
+
+.healthScore {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: stretch;
+    width: 100%;
+    gap: 15px;
+}
+
+.scoreCard {
+    background: var(--card-bg);
+    border-radius: 10px;
+    padding: 20px;
+    box-shadow: var(--shadow);
+    flex: 1;
+}
