+++ public/css/common.css
@@ -111,13 +111,12 @@ select:focus-visible {
     margin: 0;
     justify-content: center;
     align-items: center;
-    gap: 20px;
+    gap: 10px;
 }
 
 .nav div ul li {
     padding: 10px;
     border-radius: 5px;
-    width: 25%;
 }
 
 .nav div ul li a:hover {
@@ -131,6 +130,7 @@ select:focus-visible {
     text-decoration: none;
     border-radius: 5px;
     padding: 10px;
+    width: 30%;
 }
