+++ public/html/Page2Activity.html
@@ -22,9 +22,12 @@
             </div>
             <div id="two">
                 <ul>
-                    <li><a href='../index.html'>Home</a></li>
-                    <li><a href="../html/Page2Activity.html">Activity</a></li>
-                    <li><a href="../html/Page3WeeklySummary.html">Grade</a></li>
+                    <ul>
+                        <li><a href='../index.html'>Home</a></li>
+                        <li><a href="../html/Page2Activity.html">Activity</a></li>
+                        <li><a href="../html/Page3WeeklySummary.html">Weekly Summary</a></li>
+                        <li><a href="../html/Page4Grade.html">Grade</a></li>
+                    </ul>
                 </ul>
             </div>
         </div>
