+++ public/html/Page3WeeklySummary.html
@@ -22,100 +22,18 @@
             </div>
             <div id="two">
                 <ul>
-                    <li><a href='../index.html'>Home</a></li>
-                    <li><a href="../html/Page2Activity.html">Activity</a></li>
-                    <li><a href="../html/Page3WeeklySummary.html">Grade</a></li>
+                    <ul>
+                        <li><a href='../index.html'>Home</a></li>
+                        <li><a href="../html/Page2Activity.html">Activity</a></li>
+                        <li><a href="../html/Page3WeeklySummary.html">Weekly Summary</a></li>
+                        <li><a href="../html/Page4Grade.html">Grade</a></li>
+                    </ul>
                 </ul>
             </div>
         </div>
 
-        <!--Nutrition Progress Section-->
         <div class="card">
-            <div class="nutrition-summary-title">
-                <h2 class="card-title">Nutrition Summary</h2>
-                <button class="get-grade-button" id="getGradeButton">Get Your Grade!</button>
-            </div>
 
-            <!--Health Score Section-->
-            <div class="healthScore">
-                <!--Health Score Display-->
-                <div class="scoreCard">
-                    <div class="summary-card points-card">
-                        <div class="points-display">
-                            <div class="points-circle" id="pointsCircle"></div>
-                            <div class="points-label" id="underPointCircle">Your Grade</div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <!--Nutrition Totals Section-->
-            <div class="summary-grid">
-                <div class="summary-card calories-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>2000-2500 kcal</div>
-                    <div class="summary-label">Your Daily Average Calories</div>
-                    <div id="totalCalories" class="summary-value">0</div>
-                    <div class="summary-label">kcals</div>
-                    <div id="caloriesComments" class="comments"></div>
-                </div>
-                <div class="summary-card fats-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>25-50 grams</div>
-                    <div class="summary-label">Your Daily Average Total Fats</div>
-                    <div id="totalFats" class="summary-value">0</div>
-                    <div class="summary-label">grams</div>
-                    <div id="fatsComments" class="comments"></div>
-                </div>
-                <div class="summary-card cholesterol-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>0-150 milligrams</div>
-                    <div class="summary-label">Your Daily Average Cholesterol</div>
-                    <div id="totalCholesterol" class="summary-value">0</div>
-                    <div class="summary-label">milligrams</div>
-                    <div id="cholesterolComments" class="comments"></div>
-                </div>
-                <div class="summary-card sodium-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>1000-2000 milligrams</div>
-                    <div class="summary-label">Your Daily Average Sodium</div>
-                    <div id="totalSodium" class="summary-value">0</div>
-                    <div class="summary-label">milligrams</div>
-                    <div id="sodiumComments" class="comments"></div>
-                </div>
-                <div class="summary-card carbs-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>200-275 grams</div>
-                    <div class="summary-label">Your Daily Average Carbs</div>
-                    <div id="totalCarbs" class="summary-value">0</div>
-                    <div class="summary-label">grams</div>
-                    <div id="carbsComments" class="comments"></div>
-                </div>
-                <div class="summary-card fiber-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>28-38 grams</div>
-                    <div class="summary-label">Your Daily Average Fiber</div>
-                    <div id="totalFiber" class="summary-value">0</div>
-                    <div class="summary-label">grams</div>
-                    <div id="fiberComments" class="comments"></div>
-                </div>
-                <div class="summary-card sugars-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>0-30 grams</div>
-                    <div class="summary-label">Your Daily Average Sugars</div>
-                    <div id="totalSugars" class="summary-value">0</div>
-                    <div class="summary-label">grams</div>
-                    <div id="sugarsComments" class="comments"></div>
-                </div>
-                <div class="summary-card protein-card">
-                    <div class="summary-label">Daily Target</div>
-                    <div>50-75 grams</div>
-                    <div class="summary-label">Your Daily Average Protein</div>
-                    <div id="totalProtein" class="summary-value">0</div>
-                    <div class="summary-label">grams</div>
-                    <div id="proteinComments" class="comments"></div>
-                </div>
-            </div>
         </div>
     </div>
